21-01-24
Given an array of integers nums and a target integer target, write a method called subarraySum that finds the indices of a contiguous subarray in nums that add up to the target sum using a hash table (HashMap).

Your function should take two arguments:

nums: an array of integers representing the input array

target: an integer representing the target sum


Your function should return a list of two integers representing the starting and ending indices of the subarray that adds up to the target sum. If there is no such subarray, your function should return an empty list.